pyproject.toml
[project]
name = "pathway"
version = "0.1.0"
description = "Learning-aware journey state model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "ulid-py>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[project.scripts]
pathway = "pathway.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["pathway/tests"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
